Implemented TamaPet and changed all types to double to accomodate functionality for TamaPet

+++ src/main/model/TamaFood.java
@@ -6,12 +6,12 @@ package model;
 public class TamaFood {
 
     private String name;
-    private float happienessValue;
-    private float nutritionalValue;
+    private double happienessValue;
+    private double nutritionalValue;
 
     //Requires: That name is not null
     //Effects: Creates an instance of the TamaFood object
-    public TamaFood(String name, float happienessValue, float nutritionalValue) {
+    public TamaFood(String name, double happienessValue, double nutritionalValue) {
         this.name = name;
         this.happienessValue = happienessValue;
         this.nutritionalValue = nutritionalValue;
@@ -23,12 +23,12 @@ public class TamaFood {
     }
 
     //Effects: returns the happieness value of the TamaFood object
-    public float getHappieness() {
+    public double getHappieness() {
         return this.happienessValue;
     }
 
     //Effects: returns the nutritional value of the TamaFood object
-    public float getNutrition() {
+    public double getNutrition() {
         return this.nutritionalValue;
     }
 }
